requirements.txt
allure-python-commons==2.13.5
Appium-Python-Client==2.11.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
exceptiongroup==1.2.2
h11==0.14.0
idna==3.8
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
Pygments==2.18.0
PySocks==1.7.1
pytest==8.3.3
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.1.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
